requirements.txt
pandas
numpy
scikit-learn
joblib
flask
flask-cors
PyPDF2
pytesseract
Pillow
gunicorn
